Rework project structure, fix serious build weirdness
