--- TestBed/ApplicationInsights.config
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ApplicationInsights xmlns = "http://schemas.microsoft.com/ApplicationInsights/2013/Settings">
-</ApplicationInsights>
