+++ UwpReusables.Controls/Themes/Generic.xaml
@@ -1,0 +1,10 @@
+﻿<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="ms-appx:///UwpReusables.Controls/Portable/Validation/ValidatingTextBoxPortableStyle.xaml"/>
+        <ResourceDictionary Source="ms-appx:///UwpReusables.Controls/Standard/Validation/ValidatingTextBoxStyle.xaml"/>
+    </ResourceDictionary.MergedDictionaries>
+
+</ResourceDictionary>
