+++ UwpReusables.TestBed/App.xaml.cs
@@ -15,7 +15,7 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-namespace TestBed
+namespace UwpReusables.TestBed
 {
     /// <summary>
     /// Provides application-specific behavior to supplement the default Application class.
@@ -27,7 +27,7 @@ namespace TestBed
         /// executed, and as such is the logical equivalent of main() or WinMain().
         /// </summary>
         public App()
-        {            
+        {
             this.InitializeComponent();
             this.Suspending += OnSuspending;
         }
@@ -39,6 +39,14 @@ namespace TestBed
         /// <param name="e">Details about the launch request and process.</param>
         protected override void OnLaunched(LaunchActivatedEventArgs e)
         {
+
+#if DEBUG
+            if (System.Diagnostics.Debugger.IsAttached)
+            {
+                this.DebugSettings.EnableFrameRateCounter = true;
+            }
+#endif
+
             Frame rootFrame = Window.Current.Content as Frame;
 
             // Do not repeat app initialization when the Window already has content,
