+++ UwpReusables.TestBed/Assets/SplashScreen.scale-200.png
