+++ UwpReusables.TestBed/Converters/BooleanToGlyphConverter.cs
@@ -1,16 +1,14 @@
 ﻿using System;
-using Windows.UI;
 using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Media;
 
-namespace TestBed.Converters
+namespace UwpReusables.TestBed.Converters
 {
-    public class BooleanToColorConverter : IValueConverter
+    public class BooleanToGlyphConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, string language)
         {
-            bool input = (bool) value;
-            return input ? new SolidColorBrush(Colors.Green) : new SolidColorBrush(Colors.Red);
+            bool input = (bool)value;
+            return input ? "✓" : "X";
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, string language)
