+++ UwpReusables.TestBed/Package.appxmanifest
@@ -7,14 +7,14 @@
   IgnorableNamespaces="uap mp">
 
   <Identity
-    Name="44372008-f9e8-4fbc-9809-ef035e27ed2f"
+    Name="a6d4b7e8-4073-446d-9d24-29676cdc4d3f"
     Publisher="CN=nmca"
     Version="1.0.0.0" />
 
-  <mp:PhoneIdentity PhoneProductId="44372008-f9e8-4fbc-9809-ef035e27ed2f" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
+  <mp:PhoneIdentity PhoneProductId="a6d4b7e8-4073-446d-9d24-29676cdc4d3f" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
 
   <Properties>
-    <DisplayName>TestBed</DisplayName>
+    <DisplayName>UwpReusables.TestBed</DisplayName>
     <PublisherDisplayName>nmca</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
@@ -30,12 +30,12 @@
   <Applications>
     <Application Id="App"
       Executable="$targetnametoken$.exe"
-      EntryPoint="TestBed.App">
+      EntryPoint="UwpReusables.TestBed.App">
       <uap:VisualElements
-        DisplayName="TestBed"
+        DisplayName="UwpReusables.TestBed"
         Square150x150Logo="Assets\Square150x150Logo.png"
         Square44x44Logo="Assets\Square44x44Logo.png"
-        Description="TestBed"
+        Description="UwpReusables.TestBed"
         BackgroundColor="transparent">
         <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png"/>
         <uap:SplashScreen Image="Assets\SplashScreen.png" />
