+++ UwpReusables.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Controls", "Controls\Controls.csproj", "{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UwpReusables.Controls", "UwpReusables.Controls\UwpReusables.Controls.csproj", "{238B799B-7FEA-42D2-B745-90F2758F93E7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestBed", "TestBed\TestBed.csproj", "{3C4496BA-20A1-46C4-8186-C1E10667C289}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UwpReusables.TestBed", "UwpReusables.TestBed\UwpReusables.TestBed.csproj", "{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,44 +19,44 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Debug|ARM.ActiveCfg = Debug|ARM
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Debug|ARM.Build.0 = Debug|ARM
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Debug|x64.ActiveCfg = Debug|x64
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Debug|x64.Build.0 = Debug|x64
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Debug|x86.ActiveCfg = Debug|x86
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Debug|x86.Build.0 = Debug|x86
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Release|ARM.ActiveCfg = Release|ARM
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Release|ARM.Build.0 = Release|ARM
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Release|x64.ActiveCfg = Release|x64
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Release|x64.Build.0 = Release|x64
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Release|x86.ActiveCfg = Release|x86
-		{1065EA9F-DDF5-4202-B4FB-BCF6A6A55EE1}.Release|x86.Build.0 = Release|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|Any CPU.Build.0 = Debug|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|Any CPU.Deploy.0 = Debug|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|ARM.ActiveCfg = Debug|ARM
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|ARM.Build.0 = Debug|ARM
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|ARM.Deploy.0 = Debug|ARM
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|x64.ActiveCfg = Debug|x64
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|x64.Build.0 = Debug|x64
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|x64.Deploy.0 = Debug|x64
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|x86.ActiveCfg = Debug|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|x86.Build.0 = Debug|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Debug|x86.Deploy.0 = Debug|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|Any CPU.ActiveCfg = Release|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|ARM.ActiveCfg = Release|ARM
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|ARM.Build.0 = Release|ARM
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|ARM.Deploy.0 = Release|ARM
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|x64.ActiveCfg = Release|x64
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|x64.Build.0 = Release|x64
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|x64.Deploy.0 = Release|x64
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|x86.ActiveCfg = Release|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|x86.Build.0 = Release|x86
-		{3C4496BA-20A1-46C4-8186-C1E10667C289}.Release|x86.Deploy.0 = Release|x86
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Debug|ARM.ActiveCfg = Debug|ARM
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Debug|ARM.Build.0 = Debug|ARM
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Debug|x64.ActiveCfg = Debug|x64
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Debug|x64.Build.0 = Debug|x64
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Debug|x86.ActiveCfg = Debug|x86
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Debug|x86.Build.0 = Debug|x86
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Release|ARM.ActiveCfg = Release|ARM
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Release|ARM.Build.0 = Release|ARM
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Release|x64.ActiveCfg = Release|x64
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Release|x64.Build.0 = Release|x64
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Release|x86.ActiveCfg = Release|x86
+		{238B799B-7FEA-42D2-B745-90F2758F93E7}.Release|x86.Build.0 = Release|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|Any CPU.Build.0 = Debug|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|Any CPU.Deploy.0 = Debug|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|ARM.ActiveCfg = Debug|ARM
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|ARM.Build.0 = Debug|ARM
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|ARM.Deploy.0 = Debug|ARM
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|x64.ActiveCfg = Debug|x64
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|x64.Build.0 = Debug|x64
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|x64.Deploy.0 = Debug|x64
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|x86.ActiveCfg = Debug|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|x86.Build.0 = Debug|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Debug|x86.Deploy.0 = Debug|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|Any CPU.ActiveCfg = Release|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|ARM.ActiveCfg = Release|ARM
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|ARM.Build.0 = Release|ARM
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|ARM.Deploy.0 = Release|ARM
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|x64.ActiveCfg = Release|x64
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|x64.Build.0 = Release|x64
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|x64.Deploy.0 = Release|x64
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|x86.ActiveCfg = Release|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|x86.Build.0 = Release|x86
+		{25CA3AD3-2AEF-49F6-811B-1946D5447EDF}.Release|x86.Deploy.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
